Refactor new_gen creation
